+++ internal/service/blob/server.go
@@ -1,0 +1,10 @@
+package blob
+
+import pb "gitlab.com/gitlab-org/gitaly-proto/go"
+
+type server struct{}
+
+// NewServer creates a new instance of a grpc BlobServer
+func NewServer() pb.BlobServiceServer {
+	return &server{}
+}